Separate self-consistency steps on imaginary axis
Previously, the self-energy was not updated as a whole, but for each band and
frequency separately. This reduced the number of steps until convergence, but
came at the cost of losing the direct correspondence to the Dyson equation or
geometric series of Feynman diagrams. (In the superconducting state, this did
not matter, as perturbation theory breaks down anyway and the only reasonable
solution is the self-consistent one.) Also, the fixed-point iterations of the
self-energy and chemical potential were intertwined, which sometimes hindered
convergence.

Now, each step is a separate diagram evaluation. In the normal state, this is
consistent with the real-axis implementation and allows for G0W0 calculations.
The chemical potential is updated after each step.